+++ examples/main/shared/editor/EditorPlugins.tsx
@@ -105,7 +105,7 @@ import { SocialPollsServiceMock } from '../../src/Components/SocialPollsServiceM
 import 'ricos-editor/dist/styles.min.css';
 import 'wix-rich-content-plugin-commons/dist/styles.min.css';
 import 'wix-rich-content-plugin-button/dist/styles.min.css';
-// import 'wix-rich-content-plugin-code-block/dist/styles.min.css';
+import 'wix-rich-content-plugin-code-block/dist/styles.min.css';
 import 'wix-rich-content-plugin-divider/dist/styles.min.css';
 import 'wix-rich-content-plugin-emoji/dist/styles.min.css';
 import 'wix-rich-content-plugin-html/dist/styles.min.css';
@@ -551,23 +551,24 @@ const config: RichContentEditorProps['config'] = {
       new Promise(resolve =>
         setTimeout(
           () =>
-            resolve([
-              { name: searchQuery, slug: searchQuery },
-              { name: 'Test One', slug: 'testone' },
-              { name: 'Test One.1', slug: 'testone1' },
-              { name: 'Test One.2', slug: 'testone2' },
-              { name: 'Test One.3', slug: 'testone3' },
-              { name: 'Test One.4', slug: 'testone4' },
-              { name: 'Test One.5', slug: 'testone5' },
-              { name: 'Test One.6', slug: 'testone6' },
-              { name: 'Test One.7', slug: 'testone7' },
-              { name: 'Test One.8', slug: 'testone8' },
-              {
-                name: 'Test Two',
-                slug: 'testwo',
-                avatar: 'https://via.placeholder.com/100x100?text=Image=50',
-              },
-            ]),
+            resolve(
+              [
+                { name: 'Test One', slug: 'testone' },
+                { name: 'Test One.1', slug: 'testone1' },
+                { name: 'Test One.2', slug: 'testone2' },
+                { name: 'Test One.3', slug: 'testone3' },
+                { name: 'Test One.4', slug: 'testone4' },
+                { name: 'Test One.5', slug: 'testone5' },
+                { name: 'Test One.6', slug: 'testone6' },
+                { name: 'Test One.7', slug: 'testone7' },
+                { name: 'Test One.8', slug: 'testone8' },
+                {
+                  name: 'Test Two',
+                  slug: 'testwo',
+                  avatar: 'https://via.placeholder.com/100x100?text=Image=50',
+                },
+              ].filter(({ name }) => name.toLowerCase().includes(searchQuery.toLowerCase()))
+            ),
           250
         )
       ),
@@ -869,7 +870,7 @@ const config: RichContentEditorProps['config'] = {
           ],
         },
       }),
-      // buttonsMapper: () => ({
+      // buttonsOverrides: () => ({
       //   desktop: {
       //     Bold: {
       //       icon: () => <div>bla</div>,