+++ packages/editor/web/src/RichContentEditor/utils/editorCommandsUtils.ts
@@ -73,6 +73,16 @@ export const updateDocumentStyle = (editorState: EditorState, documentStyle: Doc
 
 const getInlineStylesByType = (editorState: EditorState, type: CustomInlineStyleType) => {
   const styleParser = dynamicStyleParsers[type];
+  const selection = editorState.getSelection();
+  if (selection.isCollapsed()) {
+    const currentStyles = editorState.getCurrentInlineStyle();
+    const currentStylesAsArray = currentStyles.toJS();
+    let styleToReturn;
+    currentStylesAsArray.forEach(style => {
+      styleToReturn = styleParser?.(style) ? styleParser?.(style) : styleToReturn;
+    });
+    return [styleToReturn];
+  }
   return getSelectionStyles(editorState, styleParser).map(style => styleParser(style));
 };
 
@@ -84,7 +94,9 @@ const TYPE_TO_CSS_PROPERTY = {
 
 const hasTextInSelection = (block: ContentBlock, editorState: EditorState) => {
   const blockSelectionRange = getSelectionRange(editorState, block);
-  return blockSelectionRange[0] !== blockSelectionRange[1];
+  return (
+    blockSelectionRange[0] !== blockSelectionRange[1] || editorState.getSelection().isCollapsed()
+  );
 };
 
 const getSelectionStylesFromDOM = (editorState: EditorState, type: CustomInlineStyleType) => {
@@ -123,10 +135,16 @@ const setInlineStyleByType = (
 };
 
 export const getFontSize = (editorState: EditorState) => {
-  const currentFontSizes = uniq([
-    ...getInlineStylesByType(editorState, RICOS_FONT_SIZE_TYPE),
-    ...getSelectionStylesFromDOM(editorState, RICOS_FONT_SIZE_TYPE),
-  ]);
+  const inlineFontSizes = getInlineStylesByType(editorState, RICOS_FONT_SIZE_TYPE);
+  const shouldGetStylesFromDOM = !editorState.getSelection().isCollapsed() || !inlineFontSizes?.[0];
+  const currentFontSizes = uniq(
+    [
+      ...inlineFontSizes,
+      ...(shouldGetStylesFromDOM
+        ? getSelectionStylesFromDOM(editorState, RICOS_FONT_SIZE_TYPE)
+        : []),
+    ].filter(fontSize => fontSize)
+  );
   return currentFontSizes.length === 1 ? currentFontSizes[0] : '';
 };