+++ packages/plugin-gallery/web/src/components/gallery-settings-modal.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { mergeStyles } from 'wix-rich-content-common';
+import { mergeStyles, GlobalContext } from 'wix-rich-content-common';
 import {
   Tabs,
   Tab,
@@ -10,12 +10,14 @@ import {
   FocusManager,
   SettingsMobileHeader,
   SettingsSeparator,
+  SettingsPanelHeader,
 } from 'wix-rich-content-ui-components';
 import LayoutSelector from './gallery-controls/layouts-selector';
 import styles from '../../statics/styles/gallery-settings-modal.scss';
 import LayoutControlsSection from './layout-controls-section';
 import { SortableComponent } from './gallery-controls/gallery-items-sortable';
 import { layoutData } from '../layout-data-provider';
+import classNames from 'classnames';
 const DIVIDER = 'divider';
 class ManageMediaSection extends Component {
   applyItems = items => {
@@ -26,6 +28,8 @@ class ManageMediaSection extends Component {
     };
     store.set('componentData', componentData);
   };
+
+  static contextType = GlobalContext;
 
   handleFileChange = (files, itemPos) => {
     if (files.length > 0) {
@@ -40,18 +44,10 @@ class ManageMediaSection extends Component {
   };
 
   render() {
-    const {
-      helpers,
-      store,
-      t,
-      relValue,
-      anchorTarget,
-      isMobile,
-      uiSettings,
-      languageDir,
-      accept,
-    } = this.props;
+    const { helpers, store, t, relValue, anchorTarget, isMobile, uiSettings, accept } = this.props;
     const { handleFileSelection } = helpers;
+    const { languageDir } = this.context;
+
     return (
       <div dir={languageDir} className={styles.gallerySettings_tab_section}>
         <SortableComponent
@@ -98,6 +94,8 @@ class AdvancedSettingsSection extends Component {
     store.set('componentData', componentData);
   };
 
+  static contextType = GlobalContext;
+
   switchLayout = layout => {
     this.applyGallerySetting({ ...layout, ...layoutData[layout.galleryLayout] });
   };
@@ -110,17 +108,11 @@ class AdvancedSettingsSection extends Component {
   }
 
   render() {
-    const { data, store, theme, t, isMobile, languageDir } = this.props;
+    const { data, store, theme, t, isMobile } = this.props;
+    const { languageDir } = this.context;
     return (
       this.shouldRender() && (
-        <div
-          className={
-            isMobile
-              ? styles.gallerySettings_settingsContainerMobile
-              : styles.gallerySettings_tab_section
-          }
-          dir={languageDir}
-        >
+        <div className={styles.gallerySettings_tab_section} dir={languageDir}>
           <SettingsSection
             theme={theme}
             ariaProps={{ 'aria-label': 'layout selection', role: 'region' }}
@@ -140,6 +132,7 @@ class AdvancedSettingsSection extends Component {
             store={store}
             t={t}
             isMobile={isMobile}
+            languageDir={languageDir}
           />
         </div>
       )
@@ -291,6 +284,7 @@ export class GallerySettingsModal extends Component {
           store={this.props.pubsub.store}
           helpers={this.props.helpers}
           t={this.props.t}
+          languageDir={this.props.languageDir}
         />
       </Tab>
     ),
@@ -370,7 +364,13 @@ export class GallerySettingsModal extends Component {
     this.componentData = pubsub.get('componentData');
 
     return (
-      <div data-hook="settings" dir={languageDir}>
+      <div
+        data-hook="settings"
+        className={classNames(styles.gallery_scrollContainer, {
+          [styles.gallery_scrollContainer_mobile]: isMobile,
+        })}
+        dir={languageDir}
+      >
         {isMobile && (
           <SettingsMobileHeader
             theme={theme}
@@ -379,13 +379,12 @@ export class GallerySettingsModal extends Component {
             t={t}
           />
         )}
-        <FocusManager
-          focusTrapOptions={{ initialFocus: `#${activeTab}_header` }}
-          className={styles.gallerySettings}
-          dir={languageDir}
-        >
+        <FocusManager focusTrapOptions={{ initialFocus: `#${activeTab}_header` }} dir={languageDir}>
           {!isMobile && (
-            <div className={styles.gallerySettings_title}>{t('GallerySettings_Header')}</div>
+            <SettingsPanelHeader
+              title={t('GallerySettings_Header')}
+              onClose={this.revertComponentData}
+            />
           )}
           <div className={styles.gallerySettings_tabsContainer}>
             <Tabs value={activeTab} theme={this.props.theme} onTabSelected={this.onTabSelected}>