+++ packages/ricos-common/web/src/index.ts
@@ -1,5 +1,6 @@
 export * from './types';
 export * from './utils';
+export * from './biCallbacks';
 export { RicosEngine } from './RicosEngine';
 export { default as themeStrategy } from './themeStrategy/themeStrategy';
 export { default as localeStrategy } from './localeStrategy/localeStrategy';