+++ packages/ricos-content/web/src/converters/draft/fromDraft/__tests__/fromDraft.spec.ts
@@ -11,6 +11,12 @@ import unsupportedMigrated from './unsupported-blocks-and-decorations-migrated.j
 import galleryThumbnailNone from './gallery-with-thumbnail-placement-none.json';
 import galleryWithoutThumbnail from './gallery-without-thumbnail-placement.json';
 import galleryThumbnailMigrated from './gallery-with-thumbnail-placement-migrated.json';
+import noVideoData from './no-data-video.json';
+import noVideoDataMigrated from './no-data-video-migrated.json';
+import innerWithUnsupported from './inner-rce-with-unsupported-styling.json';
+import innerWithUnsupportedMigrated from './inner-rce-with-unsupported-styling-migrated.json';
+import faultyLinkValues from './faulty-link-values.json';
+import faultyLinkValuesMigrated from './faulty-link-values-migrated.json';
 import polyfills from '../../../../../../../../e2e/tests/fixtures/polyfills.json';
 import { getTextNodes } from '../getTextNodes';
 import complexRicosFixture from '../../../../../statics/json/migratedFixtures/migration-content.json';
@@ -357,5 +363,39 @@ describe('migrate from draft', () => {
         })
       ).toEqual({})
     );
+  });
+
+  it('should convert content with no data plugins correctly', () => {
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    const converted = fromDraft(noVideoData, { ignoreUnsupportedValues: true });
+    expect(
+      compare(converted, RichContent.fromJSON(noVideoDataMigrated), {
+        ignoredKeys: ['id'],
+      })
+    ).toEqual({});
+  });
+
+  it('should convert unsupported inine inner content correctly', () => {
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    const converted = fromDraft(innerWithUnsupported, { ignoreUnsupportedValues: true });
+
+    expect(
+      compare(converted, RichContent.fromJSON(innerWithUnsupportedMigrated), {
+        ignoredKeys: ['id'],
+      })
+    ).toEqual({});
+  });
+
+  it('should convert faulty link content correctly', () => {
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    const converted = fromDraft(faultyLinkValues, { ignoreUnsupportedValues: true });
+    expect(
+      compare(converted, RichContent.fromJSON(faultyLinkValuesMigrated), {
+        ignoredKeys: ['id'],
+      })
+    ).toEqual({});
   });
 });