+++ packages/ricos-editor/web/src/modals/RicosModal.tsx
@@ -1,6 +1,7 @@
 import React, { Fragment, ComponentType, Children, FunctionComponent, ReactElement } from 'react';
 import EditorModalProvider from './EditorModalProvider';
 import { RicosEditorProps } from 'ricos-common';
+import { GlobalContext, getLangDir } from 'wix-rich-content-common';
 
 const RicosModal: FunctionComponent<RicosEditorProps & {
   container?: HTMLElement;
@@ -11,15 +12,23 @@ const RicosModal: FunctionComponent<RicosEditorProps & {
     children: {
       props: { helpers = {} },
     },
+    experiments = {},
+    isMobile = false,
+    locale,
   } = props;
   const { openModal, closeModal } = helpers;
 
   if (!openModal && !closeModal) {
     ModalProvider = EditorModalProvider;
   }
+  const context = { experiments, isMobile, languageDir: getLangDir(locale) };
 
   const child = Children.only(React.cloneElement(props.children, { ...props }));
-  return <ModalProvider {...props}>{child}</ModalProvider>;
+  return (
+    <GlobalContext.Provider value={context}>
+      <ModalProvider {...props}>{child}</ModalProvider>
+    </GlobalContext.Provider>
+  );
 };
 
 export default RicosModal;