+++ packages/tiptap-editor/web/src/commands/updateNodeById.ts
@@ -11,9 +11,9 @@ declare module '@tiptap/core' {
   }
 }
 
-export const updateNodeById: RawCommands['updateNodeById'] = (id, attrs = {}) => ({ view, tr }) => {
+export const updateNodeById: RawCommands['updateNodeById'] = (id, attrs = {}) => ({ tr }) => {
   const predicate = node => node.attrs.id === id;
-  const nodeWithPos = findChildren(view.state.doc, predicate);
+  const nodeWithPos = findChildren(tr.doc, predicate);
 
   const { pos } = nodeWithPos?.[0] || {};
   if (pos) {