+++ packages/toolbars-new/web/src/Toolbar/RicosToolbar.tsx
@@ -10,6 +10,7 @@ import {
   Link_Rel,
   CustomAnchorScroll,
   AvailableExperiments,
+  OnAddPluginLink,
 } from 'wix-rich-content-common';
 import { RicosCssOverride, RicosTheme } from 'ricos-common';
 
@@ -32,6 +33,7 @@ export type linkPanelDataType = {
     customAnchorScroll?: CustomAnchorScroll;
   };
   onLinkAdd?: any;
+  onAddPluginLink?: OnAddPluginLink;
 };
 
 export type defaultLineSpacingType = {
@@ -59,7 +61,8 @@ interface RicosToolbarProps {
   defaultLineSpacing?: defaultLineSpacingType;
   getEditorContainer: () => Element;
   cssOverride?: RicosCssOverride;
-  configButtonsMapper?: any;
+  configButtonsOverrides?: any;
+  disabled?: boolean;
 }
 
 class RicosToolbar extends Component<RicosToolbarProps> {
@@ -110,7 +113,8 @@ class RicosToolbar extends Component<RicosToolbarProps> {
       defaultLineSpacing,
       getEditorContainer,
       cssOverride,
-      configButtonsMapper,
+      configButtonsOverrides,
+      disabled,
     } = this.props;
     const updatedButtons = createButtonsList({
       buttons,
@@ -122,7 +126,7 @@ class RicosToolbar extends Component<RicosToolbarProps> {
       defaultLineSpacing,
       experiments,
       theme,
-      configButtonsMapper,
+      configButtonsOverrides,
     });
     const buttonsWithoutUnwantedSeparators =
       updatedButtons.length > 0 && this.cleanUnwantedSeparators(updatedButtons);
@@ -140,6 +144,7 @@ class RicosToolbar extends Component<RicosToolbarProps> {
         nestedMenu={nestedMenu}
         onToolbarButtonClick={this.props.onToolbarButtonClick}
         getEditorContainer={getEditorContainer}
+        disabled={disabled}
       />
     );
   }