+++ packages/viewer/web/src/RichContentViewer.tsx
@@ -165,12 +165,11 @@ class RichContentViewer extends Component<
       seoMode,
       iframeSandboxDomain,
       textAlignment,
+      experiments,
     }: RichContentViewerProps,
     contentState?: DraftContent
   ): ViewerContextType => {
     deprecateHelpers(helpers, config);
-    const version = Version.currentVersion;
-    const contentId = this.props.initialState?.ID;
     return {
       t,
       theme,
@@ -179,12 +178,7 @@ class RichContentViewer extends Component<
       relValue,
       customAnchorScroll,
       config,
-      helpers: {
-        ...helpers,
-        onViewerLoaded: args => helpers.onViewerLoaded?.({ ...args, version, contentId }),
-        onViewerAction: (pluginId, actionName, value) =>
-          helpers.onViewerAction?.(pluginId, actionName, value, contentId),
-      },
+      helpers,
       locale,
       disabled,
       seoMode,
@@ -192,6 +186,7 @@ class RichContentViewer extends Component<
       iframeSandboxDomain,
       disableRightClick: config?.uiSettings?.disableRightClick,
       textAlignment,
+      experiments,
     };
   };
 
@@ -251,6 +246,7 @@ class RichContentViewer extends Component<
       typeMappers,
       setRef = () => {},
       onMouseOver = () => {},
+      experiments = {},
     } = this.props;
     const decorators = [...this.props.decorators, createJustificationFixDecorator()];
     try {
@@ -267,8 +263,10 @@ class RichContentViewer extends Component<
         viewerStyles.renderedInTable,
         draftDefaultStyles.renderedInTable
       );
+
       const editorClassName = classNames(styles.editor, renderedInTable && tableClassNames, {
         [styles.rtl]: textDirection === 'rtl',
+        [styles.fixedTabSize]: experiments.fixedTabSize?.enabled,
       });
 
       const initSpoilers = config[SPOILER_TYPE]?.initSpoilersContentState;
@@ -294,7 +292,7 @@ class RichContentViewer extends Component<
         config,
         initSpoilers,
         SpoilerViewerWrapper,
-        { addAnchors },
+        { addAnchors, fixedTabSize: experiments.fixedTabSize?.enabled },
         innerRCEViewerProps,
         this.props.documentStyle
       );